+++ py_bootstrap/operations/export_bootstrap.py
@@ -55,50 +55,37 @@ class ExportBootstrapsDispatcherOperation(BaseBootstrapsOperation):
         entry_points_modules_map = dict(self.find_bootstraps())
 
         entry_point_module = entry_points_modules_map[bootstrap_name]
+        assert entry_point_module.__file__
 
-        try:
-            operation: "BaseExportBootstrapOperation" = (
-                entry_point_module.ExportOperation()
-            )
-        except AttributeError as err:
-            logger.error(
-                "%r. metadata does not contain BuildOperation attribute: %r.",
-                self,
-                entry_point_module,
-            )
-            raise Exception(
-                f"Invalid metadata module for {bootstrap_name}"
-            ) from err
-        except Exception as err:
-            logger.exception(
-                "%r. unable to create operation instance from: %r. error: %r.",
-                self,
-                entry_point_module.Operation,
-                err,
-            )
-            raise Exception(
-                f"BuildOperation creation failed for {bootstrap_name}"
-            ) from err
+        operation: "BaseExportBootstrapOperation" = (
+            entry_point_module.ExportOperation()
+        )
 
-        operation.set_cli_namespace(self.cli_namespace)
+        operation.set_cli_namespace(namespace=self.cli_namespace)
+        operation.set_bootstrap_path(
+            path=Path(entry_point_module.__file__).parent
+        )
         operation.run()
 
 
 class BaseExportBootstrapOperation(BaseBootstrapsOperation):
-    entry_point_path: t.ClassVar[str]
+    _bootstrap_path: "Path"
 
     @classmethod
     def prepare_cli_parser(cls, parser: "ArgumentParser", prefix: str = ""): ...
 
     @cached_property
     def bootstrap_path(self) -> "Path":
-        return Path(self.entry_point_path).parent
+        return self._bootstrap_path
 
     @cached_property
     def destination_path(self) -> "Path":
         if self.cli_namespace.destination_dir:
             return Path(os.getcwd(), self._cli_namespace.destination_dir)
         return Path.cwd()
+
+    def set_bootstrap_path(self, path: "Path"):
+        self._bootstrap_path = path
 
     def run(self):
         self.create_destination_dir()