+++ py_bootstrap/templates/__init__.py
@@ -1,0 +1,7 @@
+__all__ = ("ENABLED_TEMPLATES",)
+
+ENABLED_TEMPLATES = [
+    "application",
+    "package",
+    "bootstrap",
+]