+++ py_bootstrap/templates/package/__entry_point__.py
@@ -23,7 +23,6 @@ class BuildOperation(BaseBuildBootstrapOperation):
     cli_description = "Generates a skeleton of a new Python Package"
     cli_argument_name_help = "Specifies name of the package"
     cli_argument_description_help = "Specifies description of the package"
-    entry_point_path = __file__
 
     @classmethod
     def prepare_cli_parser(cls, parser: "ArgumentParser", prefix: str = ""):
@@ -62,4 +61,3 @@ class BuildOperation(BaseBuildBootstrapOperation):
 
 class ExportOperation(BaseExportBootstrapOperation):
     cli_description = "Exports a Python Package's template files"
-    entry_point_path = __file__