+++ tests/tst_templates/__init__.py
@@ -1,0 +1,7 @@
+__all__ = ("ENABLED_TEMPLATES",)
+
+ENABLED_TEMPLATES = [
+    "test_bootstrap",
+    "test_missed_entry_point",
+    "test_wrong_entry_point",
+]