+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### [0.6.1] - 2025-06-24
+## [0.7.0] - 2025-06-24
+### Changed
+- Convert `templates` directory into a Python package. See `py_bootstrap/templates/__init__.py` file for details.
+- Implement getting a list of bootstraps from setuptools plugins. See `py_bootstrap/operations/base.py` file for details.
+
+## [0.6.1] - 2025-06-24
 ### Changed
 - Refactor `README.md` files in `application`, `package`. See `py_bootastrap/templates/` directory for details.
 - Refactor project's `README.md` file. See `README.md` for details.